Pull page-level AV config out into AudioVideoHead component & use once in page
